\name{tREXopt-package}
\alias{tREXopt-package}
\alias{tREXopt}
\docType{package}
\title{
\packageTitle{tREXopt}
}
\description{
\packageDescription{tREXopt}
}
\details{
\packageIndices{tREXopt}
tREX-opt: A Random Effect Optimization Method for Recapitulation of Single Cell Chromatin 3D Structure.
}
\author{
\packageAuthor{tREXopt}
Maintainer: \packageMaintainer{tREXopt}
}
\references{
Park, J. and Lin, S. (2016). Impact of data resolution on three-dimensional structure inference methods. BMC bioinformatics, 17(1), 70.
}

